blorgo/pyproject.toml
[project]
name = "blorgo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "resonate-sdk>=0.5.3",
    "torch>=2.8.0",
    "transformers>=4.55.0",
]
